Move `URLProtocolStub` class to separate file
